+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:travel_muse_app/views/my_page/edit_profile_sheet.dart';
+import 'package:travel_muse_app/views/my_page/sheet/edit_profile_sheet.dart';
 import 'package:travel_muse_app/views/login/login_page.dart';
 
 void main() {
