+++ lib/views/my_page/sheet/widgets/edit_profile_image.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+class EditProfileImage extends StatelessWidget {
+  const EditProfileImage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        // 이미지피커
+      },
+      child: SizedBox(
+        width: 120,
+        height: 120,
+        child: Stack(
+          children: [
+            ClipRRect(
+              borderRadius: BorderRadius.circular(100),
+              // 추후 파이어베이스 이미지 url연결
+              child: Image.network(
+                'https://picsum.photos/id/1/200/200',
+                fit: BoxFit.cover,
+              ),
+            ),
+            Align(
+              alignment: Alignment.bottomRight,
+              child: Container(
+                width: 40,
+                height: 40,
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(50),
+                  color: Colors.grey[200],
+                  border: Border.all(width: 1, color: Colors.grey[400]!),
+                ),
+                child: Icon(
+                  Icons.camera_alt,
+                  color: Colors.grey[600],
+                  size: 28,
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
